@@ -4,7 +4,7 @@ version: 2.0.0-dev.2
 homepage: https://github.com/Albert-Jan/avataaar-image-2/
 
 environment:
-  sdk: ">=2.2.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
