@@ -4,10 +4,10 @@ import 'package:cached_network_image/cached_network_image.dart';
 import 'package:flutter/widgets.dart';
 
 class AvataaarImage extends StatelessWidget {
-  AvataaarImage({
-    Key key,
-    @required this.baseUrl,
-    @required this.avatar,
+  const AvataaarImage({
+    Key? key,
+    required this.baseUrl,
+    required this.avatar,
     this.width: 64.0,
     this.placeholder,
     this.errorImage,
@@ -15,31 +15,32 @@ class AvataaarImage extends StatelessWidget {
         super(key: key);
 
   AvataaarImage.builder({
-    Key key,
-    @required this.builder,
-    @required this.baseUrl,
-    @required this.avatar,
+    Key? key,
+    required this.builder,
+    required this.baseUrl,
+    required this.avatar,
     this.width: 64.0,
   })  : this.placeholder = null,
         this.errorImage = null,
         super(key: key);
 
   final String baseUrl;
-  final Widget Function(BuildContext context, String url) builder;
+  final Widget Function(BuildContext context, String url)? builder;
   final Avataaar avatar;
-  final Widget placeholder;
-  final Widget errorImage;
+  final Widget? placeholder;
+  final Widget? errorImage;
   final double width;
 
   @override
   Widget build(BuildContext context) {
     final imageUrl = AvataaarsApi().getUrl(baseUrl, avatar, width);
     return builder != null
-        ? builder(context, imageUrl)
+        ? builder!(context, imageUrl)
         : CachedNetworkImage(
             imageUrl: imageUrl,
-            placeholder: placeholder != null ? (_, __) => placeholder : null,
-            errorWidget: errorImage != null ? (_, __, ___) => errorImage : null,
+            placeholder: placeholder != null ? (_, __) => placeholder! : null,
+            errorWidget:
+                errorImage != null ? (_, __, ___) => errorImage! : null,
           );
   }
 }
