@@ -3,9 +3,9 @@ import 'package:avataaar_image_2/src/parts/parts.dart';
 import 'package:avataaar_image_2/src/pieces.dart';
 
 class Eyebrow implements AvataaarPart {
-  Eyebrow._({this.eyebrowType});
+  const Eyebrow._({this.eyebrowType});
 
-  final EyebrowType eyebrowType;
+  final EyebrowType? eyebrowType;
 
   @override
   List get pieces => [eyebrowType];
@@ -60,7 +60,7 @@ class Eyebrow implements AvataaarPart {
 
 class EyebrowConverter extends Converter<Eyebrow> {
   @override
-  Eyebrow fromMap(Map<String, dynamic> map) {
+  Eyebrow? fromMap(Map<String, dynamic>? map) {
     if (map == null) return null;
     return Eyebrow._(
       eyebrowType: enumFromJson(EyebrowType.values, map['eyebrowType']),
@@ -68,7 +68,7 @@ class EyebrowConverter extends Converter<Eyebrow> {
   }
 
   @override
-  Map<String, dynamic> toMap(Eyebrow value) {
+  Map<String, dynamic>? toMap(Eyebrow? value) {
     if (value == null) return null;
     return {'eyebrowType': enumToJson(value.eyebrowType)};
   }
