@@ -4,7 +4,7 @@ import 'package:avataaar_image_2/src/parts/parts.dart';
 import 'package:avataaar_image_2/src/pieces.dart';
 
 class Top implements AvataaarPart {
-  Top._({
+  const Top._({
     this.topType,
     this.accessoriesType,
     this.hatColor,
@@ -12,11 +12,11 @@ class Top implements AvataaarPart {
     this.facialHair,
   });
 
-  final TopType topType;
-  final AccessoriesType accessoriesType;
-  final HatColor hatColor;
-  final HairColor hairColor;
-  final FacialHair facialHair;
+  final TopType? topType;
+  final AccessoriesType? accessoriesType;
+  final HatColor? hatColor;
+  final HairColor? hairColor;
+  final FacialHair? facialHair;
 
   @override
   List get pieces => [
@@ -45,35 +45,36 @@ class Top implements AvataaarPart {
       hairColor.hashCode ^
       facialHair.hashCode;
 
-  static Top noHair({AccessoriesType accessoriesType, FacialHair facialHair}) =>
+  static Top noHair(
+          {AccessoriesType? accessoriesType, FacialHair? facialHair}) =>
       Top._(
           topType: TopType.NoHair,
           accessoriesType: accessoriesType,
           facialHair: facialHair);
 
   static Top eyepatch(
-          {AccessoriesType accessoriesType, FacialHair facialHair}) =>
+          {AccessoriesType? accessoriesType, FacialHair? facialHair}) =>
       Top._(
           topType: TopType.Eyepatch,
           accessoriesType: accessoriesType,
           facialHair: facialHair);
 
-  static Top hat({AccessoriesType accessoriesType, FacialHair facialHair}) =>
+  static Top hat({AccessoriesType? accessoriesType, FacialHair? facialHair}) =>
       Top._(
           topType: TopType.Hat,
           accessoriesType: accessoriesType,
           facialHair: facialHair);
 
-  static Top hijab({AccessoriesType accessoriesType, HatColor hatColor}) =>
+  static Top hijab({AccessoriesType? accessoriesType, HatColor? hatColor}) =>
       Top._(
           topType: TopType.Hijab,
           accessoriesType: accessoriesType,
           hatColor: hatColor);
 
   static Top turban(
-          {AccessoriesType accessoriesType,
-          HatColor hatColor,
-          FacialHair facialHair}) =>
+          {AccessoriesType? accessoriesType,
+          HatColor? hatColor,
+          FacialHair? facialHair}) =>
       Top._(
           topType: TopType.Turban,
           accessoriesType: accessoriesType,
@@ -81,9 +82,9 @@ class Top implements AvataaarPart {
           facialHair: facialHair);
 
   static Top winterHat1(
-          {AccessoriesType accessoriesType,
-          HatColor hatColor,
-          FacialHair facialHair}) =>
+          {AccessoriesType? accessoriesType,
+          HatColor? hatColor,
+          FacialHair? facialHair}) =>
       Top._(
           topType: TopType.WinterHat1,
           accessoriesType: accessoriesType,
@@ -91,9 +92,9 @@ class Top implements AvataaarPart {
           facialHair: facialHair);
 
   static Top winterHat2(
-          {AccessoriesType accessoriesType,
-          HatColor hatColor,
-          FacialHair facialHair}) =>
+          {AccessoriesType? accessoriesType,
+          HatColor? hatColor,
+          FacialHair? facialHair}) =>
       Top._(
           topType: TopType.WinterHat2,
           accessoriesType: accessoriesType,
@@ -101,9 +102,9 @@ class Top implements AvataaarPart {
           facialHair: facialHair);
 
   static Top winterHat3(
-          {AccessoriesType accessoriesType,
-          HatColor hatColor,
-          FacialHair facialHair}) =>
+          {AccessoriesType? accessoriesType,
+          HatColor? hatColor,
+          FacialHair? facialHair}) =>
       Top._(
           topType: TopType.WinterHat3,
           accessoriesType: accessoriesType,
@@ -111,9 +112,9 @@ class Top implements AvataaarPart {
           facialHair: facialHair);
 
   static Top winterHat4(
-          {AccessoriesType accessoriesType,
-          HatColor hatColor,
-          FacialHair facialHair}) =>
+          {AccessoriesType? accessoriesType,
+          HatColor? hatColor,
+          FacialHair? facialHair}) =>
       Top._(
           topType: TopType.WinterHat4,
           accessoriesType: accessoriesType,
@@ -121,9 +122,9 @@ class Top implements AvataaarPart {
           facialHair: facialHair);
 
   static Top longHairBigHair(
-          {AccessoriesType accessoriesType,
-          HairColor hairColor,
-          FacialHair facialHair}) =>
+          {AccessoriesType? accessoriesType,
+          HairColor? hairColor,
+          FacialHair? facialHair}) =>
       Top._(
           topType: TopType.LongHairBigHair,
           accessoriesType: accessoriesType,
@@ -131,9 +132,9 @@ class Top implements AvataaarPart {
           facialHair: facialHair);
 
   static Top longHairBob(
-          {AccessoriesType accessoriesType,
-          HairColor hairColor,
-          FacialHair facialHair}) =>
+          {AccessoriesType? accessoriesType,
+          HairColor? hairColor,
+          FacialHair? facialHair}) =>
       Top._(
           topType: TopType.LongHairBob,
           accessoriesType: accessoriesType,
@@ -141,9 +142,9 @@ class Top implements AvataaarPart {
           facialHair: facialHair);
 
   static Top longHairBun(
-          {AccessoriesType accessoriesType,
-          HairColor hairColor,
-          FacialHair facialHair}) =>
+          {AccessoriesType? accessoriesType,
+          HairColor? hairColor,
+          FacialHair? facialHair}) =>
       Top._(
           topType: TopType.LongHairBun,
           accessoriesType: accessoriesType,
@@ -151,9 +152,9 @@ class Top implements AvataaarPart {
           facialHair: facialHair);
 
   static Top longHairCurly(
-          {AccessoriesType accessoriesType,
-          HairColor hairColor,
-          FacialHair facialHair}) =>
+          {AccessoriesType? accessoriesType,
+          HairColor? hairColor,
+          FacialHair? facialHair}) =>
       Top._(
           topType: TopType.LongHairCurly,
           accessoriesType: accessoriesType,
@@ -161,9 +162,9 @@ class Top implements AvataaarPart {
           facialHair: facialHair);
 
   static Top longHairCurvy(
-          {AccessoriesType accessoriesType,
-          HairColor hairColor,
-          FacialHair facialHair}) =>
+          {AccessoriesType? accessoriesType,
+          HairColor? hairColor,
+          FacialHair? facialHair}) =>
       Top._(
           topType: TopType.LongHairCurvy,
           accessoriesType: accessoriesType,
@@ -171,9 +172,9 @@ class Top implements AvataaarPart {
           facialHair: facialHair);
 
   static Top longHairDreads(
-          {AccessoriesType accessoriesType,
-          HairColor hairColor,
-          FacialHair facialHair}) =>
+          {AccessoriesType? accessoriesType,
+          HairColor? hairColor,
+          FacialHair? facialHair}) =>
       Top._(
           topType: TopType.LongHairDreads,
           accessoriesType: accessoriesType,
@@ -181,16 +182,16 @@ class Top implements AvataaarPart {
           facialHair: facialHair);
 
   static Top longHairFrida(
-          {AccessoriesType accessoriesType, FacialHair facialHair}) =>
+          {AccessoriesType? accessoriesType, FacialHair? facialHair}) =>
       Top._(
           topType: TopType.LongHairFrida,
           accessoriesType: accessoriesType,
           facialHair: facialHair);
 
   static Top longHairFro(
-          {AccessoriesType accessoriesType,
-          HairColor hairColor,
-          FacialHair facialHair}) =>
+          {AccessoriesType? accessoriesType,
+          HairColor? hairColor,
+          FacialHair? facialHair}) =>
       Top._(
           topType: TopType.LongHairFro,
           accessoriesType: accessoriesType,
@@ -198,9 +199,9 @@ class Top implements AvataaarPart {
           facialHair: facialHair);
 
   static Top longHairFroBand(
-          {AccessoriesType accessoriesType,
-          HairColor hairColor,
-          FacialHair facialHair}) =>
+          {AccessoriesType? accessoriesType,
+          HairColor? hairColor,
+          FacialHair? facialHair}) =>
       Top._(
           topType: TopType.LongHairFroBand,
           accessoriesType: accessoriesType,
@@ -208,9 +209,9 @@ class Top implements AvataaarPart {
           facialHair: facialHair);
 
   static Top longHairNotTooLong(
-          {AccessoriesType accessoriesType,
-          HairColor hairColor,
-          FacialHair facialHair}) =>
+          {AccessoriesType? accessoriesType,
+          HairColor? hairColor,
+          FacialHair? facialHair}) =>
       Top._(
           topType: TopType.LongHairNotTooLong,
           accessoriesType: accessoriesType,
@@ -218,16 +219,16 @@ class Top implements AvataaarPart {
           facialHair: facialHair);
 
   static Top longHairShavedSides(
-          {AccessoriesType accessoriesType, FacialHair facialHair}) =>
+          {AccessoriesType? accessoriesType, FacialHair? facialHair}) =>
       Top._(
           topType: TopType.LongHairShavedSides,
           accessoriesType: accessoriesType,
           facialHair: facialHair);
 
   static Top longHairMiaWallace(
-          {AccessoriesType accessoriesType,
-          HairColor hairColor,
-          FacialHair facialHair}) =>
+          {AccessoriesType? accessoriesType,
+          HairColor? hairColor,
+          FacialHair? facialHair}) =>
       Top._(
           topType: TopType.LongHairMiaWallace,
           accessoriesType: accessoriesType,
@@ -235,9 +236,9 @@ class Top implements AvataaarPart {
           facialHair: facialHair);
 
   static Top longHairStraight(
-          {AccessoriesType accessoriesType,
-          HairColor hairColor,
-          FacialHair facialHair}) =>
+          {AccessoriesType? accessoriesType,
+          HairColor? hairColor,
+          FacialHair? facialHair}) =>
       Top._(
           topType: TopType.LongHairStraight,
           accessoriesType: accessoriesType,
@@ -245,9 +246,9 @@ class Top implements AvataaarPart {
           facialHair: facialHair);
 
   static Top longHairStraight2(
-          {AccessoriesType accessoriesType,
-          HairColor hairColor,
-          FacialHair facialHair}) =>
+          {AccessoriesType? accessoriesType,
+          HairColor? hairColor,
+          FacialHair? facialHair}) =>
       Top._(
           topType: TopType.LongHairStraight2,
           accessoriesType: accessoriesType,
@@ -255,9 +256,9 @@ class Top implements AvataaarPart {
           facialHair: facialHair);
 
   static Top longHairStraightStrand(
-          {AccessoriesType accessoriesType,
-          HairColor hairColor,
-          FacialHair facialHair}) =>
+          {AccessoriesType? accessoriesType,
+          HairColor? hairColor,
+          FacialHair? facialHair}) =>
       Top._(
           topType: TopType.LongHairStraightStrand,
           accessoriesType: accessoriesType,
@@ -265,9 +266,9 @@ class Top implements AvataaarPart {
           facialHair: facialHair);
 
   static Top shortHairDreads01(
-          {AccessoriesType accessoriesType,
-          HairColor hairColor,
-          FacialHair facialHair}) =>
+          {AccessoriesType? accessoriesType,
+          HairColor? hairColor,
+          FacialHair? facialHair}) =>
       Top._(
           topType: TopType.ShortHairDreads01,
           accessoriesType: accessoriesType,
@@ -275,9 +276,9 @@ class Top implements AvataaarPart {
           facialHair: facialHair);
 
   static Top shortHairDreads02(
-          {AccessoriesType accessoriesType,
-          HairColor hairColor,
-          FacialHair facialHair}) =>
+          {AccessoriesType? accessoriesType,
+          HairColor? hairColor,
+          FacialHair? facialHair}) =>
       Top._(
           topType: TopType.ShortHairDreads02,
           accessoriesType: accessoriesType,
@@ -285,9 +286,9 @@ class Top implements AvataaarPart {
           facialHair: facialHair);
 
   static Top shortHairFrizzle(
-          {AccessoriesType accessoriesType,
-          HairColor hairColor,
-          FacialHair facialHair}) =>
+          {AccessoriesType? accessoriesType,
+          HairColor? hairColor,
+          FacialHair? facialHair}) =>
       Top._(
           topType: TopType.ShortHairFrizzle,
           accessoriesType: accessoriesType,
@@ -295,9 +296,9 @@ class Top implements AvataaarPart {
           facialHair: facialHair);
 
   static Top shortHairShaggyMullet(
-          {AccessoriesType accessoriesType,
-          HairColor hairColor,
-          FacialHair facialHair}) =>
+          {AccessoriesType? accessoriesType,
+          HairColor? hairColor,
+          FacialHair? facialHair}) =>
       Top._(
           topType: TopType.ShortHairShaggyMullet,
           accessoriesType: accessoriesType,
@@ -305,9 +306,9 @@ class Top implements AvataaarPart {
           facialHair: facialHair);
 
   static Top shortHairShortCurly(
-          {AccessoriesType accessoriesType,
-          HairColor hairColor,
-          FacialHair facialHair}) =>
+          {AccessoriesType? accessoriesType,
+          HairColor? hairColor,
+          FacialHair? facialHair}) =>
       Top._(
           topType: TopType.ShortHairShortCurly,
           accessoriesType: accessoriesType,
@@ -315,9 +316,9 @@ class Top implements AvataaarPart {
           facialHair: facialHair);
 
   static Top shortHairShortFlat(
-          {AccessoriesType accessoriesType,
-          HairColor hairColor,
-          FacialHair facialHair}) =>
+          {AccessoriesType? accessoriesType,
+          HairColor? hairColor,
+          FacialHair? facialHair}) =>
       Top._(
           topType: TopType.ShortHairShortFlat,
           accessoriesType: accessoriesType,
@@ -325,9 +326,9 @@ class Top implements AvataaarPart {
           facialHair: facialHair);
 
   static Top shortHairShortRound(
-          {AccessoriesType accessoriesType,
-          HairColor hairColor,
-          FacialHair facialHair}) =>
+          {AccessoriesType? accessoriesType,
+          HairColor? hairColor,
+          FacialHair? facialHair}) =>
       Top._(
           topType: TopType.ShortHairShortRound,
           accessoriesType: accessoriesType,
@@ -335,9 +336,9 @@ class Top implements AvataaarPart {
           facialHair: facialHair);
 
   static Top shortHairShortWaved(
-          {AccessoriesType accessoriesType,
-          HairColor hairColor,
-          FacialHair facialHair}) =>
+          {AccessoriesType? accessoriesType,
+          HairColor? hairColor,
+          FacialHair? facialHair}) =>
       Top._(
           topType: TopType.ShortHairShortWaved,
           accessoriesType: accessoriesType,
@@ -345,9 +346,9 @@ class Top implements AvataaarPart {
           facialHair: facialHair);
 
   static Top shortHairSides(
-          {AccessoriesType accessoriesType,
-          HairColor hairColor,
-          FacialHair facialHair}) =>
+          {AccessoriesType? accessoriesType,
+          HairColor? hairColor,
+          FacialHair? facialHair}) =>
       Top._(
           topType: TopType.ShortHairSides,
           accessoriesType: accessoriesType,
@@ -355,9 +356,9 @@ class Top implements AvataaarPart {
           facialHair: facialHair);
 
   static Top shortHairTheCaesar(
-          {AccessoriesType accessoriesType,
-          HairColor hairColor,
-          FacialHair facialHair}) =>
+          {AccessoriesType? accessoriesType,
+          HairColor? hairColor,
+          FacialHair? facialHair}) =>
       Top._(
           topType: TopType.ShortHairTheCaesar,
           accessoriesType: accessoriesType,
@@ -365,16 +366,16 @@ class Top implements AvataaarPart {
           facialHair: facialHair);
 
   static Top shortHairTheCaesarSidePart(
-          {AccessoriesType accessoriesType,
-          HairColor hairColor,
-          FacialHair facialHair}) =>
+          {AccessoriesType? accessoriesType,
+          HairColor? hairColor,
+          FacialHair? facialHair}) =>
       Top._(
           topType: TopType.ShortHairTheCaesarSidePart,
           accessoriesType: accessoriesType,
           hairColor: hairColor,
           facialHair: facialHair);
 
-  static Top get random {
+  static Top? get random {
     final topType = randomPiece(TopType.values);
     final accessoriesType = randomPiece(AccessoriesType.values);
     final hatColor = randomPiece(HatColor.values);
@@ -549,7 +550,7 @@ class Top implements AvataaarPart {
 
 class TopConverter extends Converter<Top> {
   @override
-  Top fromMap(Map<String, dynamic> map) {
+  Top? fromMap(Map<String, dynamic>? map) {
     if (map == null) return null;
     return Top._(
       topType: enumFromJson(TopType.values, map['topType']),
@@ -562,7 +563,7 @@ class TopConverter extends Converter<Top> {
   }
 
   @override
-  Map<String, dynamic> toMap(Top value) {
+  Map<String, dynamic>? toMap(Top? value) {
     if (value == null) return null;
     return {
       'topType': enumToJson(value.topType),
