@@ -7,8 +7,8 @@ import '../avataaar_image_2.dart';
 class AvataaarsApi {
   static const baseUrl = 'https://avataaars.io';
 
-  AvataaarsApi();
-  
+  const AvataaarsApi();
+
   String getUrl(String baseUrl, Avataaar avatar, double width) {
     final params = avatar.pieceEntries.map((it) {
       final key = it.key[0].toLowerCase() + it.key.substring(1);
@@ -17,6 +17,7 @@ class AvataaarsApi {
     return '$baseUrl/png/$width?$params';
   }
 
-  Future<Uint8List> getImage(Avataaar avatar, double width) =>
-      http.get(Uri(path: getUrl(baseUrl, avatar, width))).then((it) => it.bodyBytes);
+  Future<Uint8List> getImage(Avataaar avatar, double width) => http
+      .get(Uri(path: getUrl(baseUrl, avatar, width)))
+      .then((it) => it.bodyBytes);
 }
